requirements.txt
langchain==0.3.13
langchain-groq==0.2.1
langgraph==0.2.59
streamlit==1.40.2
python-dotenv==1.0.0
tavily-python==0.5.0
pydantic==2.10.3
opencv-python==4.10.0.84
streamlit-webrtc==0.47.10
pillow==11.0.0
google-generativeai==0.8.3
PyPDF2==3.0.1
pdfplumber==0.11.4
